Move a few types from vessel entity to vessel feature
