+++ frontend/src/features/Vessel/Vessel.types.ts
@@ -1,4 +1,7 @@
-import type { RiskFactor } from '../../domain/entities/vessel/riskFactor/types'
+import type { ReportingType } from '@features/Reporting/types'
+import type { PendingAlertValueType } from 'domain/entities/alerts/types'
+import type { RiskFactor } from 'domain/entities/vessel/riskFactor/types'
+import type { VesselLastPosition } from 'domain/entities/vessel/types'
 
 export namespace Vessel {
   export type Beacon = {
@@ -48,5 +51,23 @@ export namespace Vessel {
     logbookEquipmentStatus: string | undefined
     logbookSoftware: string | undefined
     riskFactor: RiskFactor | undefined
+  }
+
+  export type VesselEnhancedObject = VesselLastPosition & {
+    alerts: Array<PendingAlertValueType | 'PNO_LAN_WEIGHT_TOLERANCE_ALERT'>
+    flagState: string
+    fleetSegmentsArray: string[]
+    gearsArray: string[]
+    hasAlert: boolean
+    hasInfractionSuspicion: boolean
+    lastControlDateTimeTimestamp: number | string
+    reportings: ReportingType[]
+    speciesArray: string[]
+  }
+
+  export type SelectedVessel = VesselEnhancedObject & Vessel.EnrichedVessel
+  export type AugmentedSelectedVessel = SelectedVessel & {
+    hasAlert: boolean
+    hasInfractionSuspicion: boolean
   }
 }
